# Skip the creation of a transform stream in fetch