@mcollina 2024-04-11
From the spec:

> This [TransformStream](https://streams.spec.whatwg.org/#transformstream) is needed for the purpose of receiving a notification when the stream reaches its end, and is otherwise an [identity transform stream](https://streams.spec.whatwg.org/#identity-transform-stream).

which we fulfill in https://github.com/nodejs/undici/pull/3093/files#diff-9cc53c7d17eaefd294df4a58f7406660dfe9e2b39feea783e824d33095cafedeR1106.

---

So in other terms, it seems they had no easy way to specify what we did in our implementation.