@KhafraDev 2024-04-11
We can *almost* get rid of webstreams here entirely (only 4 WPTs fail):

```diff
diff --git a/lib/web/fetch/index.js b/lib/web/fetch/index.js
index 30f457df..21a7f0ae 100644
--- a/lib/web/fetch/index.js
+++ b/lib/web/fetch/index.js
@@ -1089,32 +1089,9 @@ function fetchFinale (fetchParams, response) {
     //    set to processResponseEndOfBody.
     // 4. Set internalResponse’s body’s stream to the result of internalResponse’s body’s stream piped through transformStream.

-    const byteStream = new ReadableStream({
-      readableStream: internalResponse.body.stream,
-      async start () {
-        this._bodyReader = this.readableStream.getReader()
-      },
-      async pull (controller) {
-        while (controller.desiredSize >= 0) {
-          const { done, value } = await this._bodyReader.read()
-
-          if (done) {
-            queueMicrotask(() => {
-              readableStreamClose(controller)
-
-              // Added to in lieu of steps 1-4 of the spec
-              processResponseEndOfBody()
-            })
-            break
-          }
-
-          controller.enqueue(value)
-        }
-      },
-      type: 'bytes'
+    require('node:stream').finished(internalResponse.body.stream, (err) => {
+      processResponseEndOfBody()
     })
-
-    internalResponse.body.stream = byteStream
   }
 }
```