@KhafraDev 2024-04-12
With the benchmarks, it was ~3.5% faster, but that's only the difference between a couple of promise allocations (from the stream) and a single one from stream.finished. The larger the body, the bigger the gains should be.